Following bug report and fix sent by Michael Chuvelev (Intel) on Fri, 2 Apr 2010 to "lapack@cs.utk.edu". Issue in LAPACK 3.2.1 *la_{po,sy,he}rfsx_extended.f sources. For instanse, in sla_syrfsx_extended.f:
DY is described as
*     DY             (input) REAL array, dimension (N)
*     Workspace to hold the intermediate solution.
 
whereas N-by-NRHS array is referenced at DY due to the following statement (line 384):
 
            CALL SSYTRS( UPLO, N, NRHS, AF, LDAF, IPIV, DY, N, INFO )
 
This may lead to an outside memory corruption. I assume one RHS should be referenced here, because all the RHSs are considered individually in the loop by J. That is, substituting NRHS by 1:
 
            CALL SSYTRS( UPLO, N, 1, AF, LDAF, IPIV, DY, N, INFO )
 
should fix the issue. This is actual for any subroutine mentioned.
 


git-svn-id: https://icl.cs.utk.edu/svn/lapack-dev/lapack/trunk@736 8a072113-8704-0410-8d35-dd094bca7971