Correct Bug in the testing of CHERFSX function reported by omitrofa (olga.mitrofanova@intel.com) See Forum 2644
Test of CHERFSX function doesn't work because there is an error
in ../TESTING/LIN/cerrhex.f ( lines 284-330 ). 
This function uses the diagonal pivoting factorization of a Hermitian
indefinite matrix instead of the diagonal pivoting factorization of
a Hermitian indefinite packed matrix. When CHERFSX is included in
conditional "IF( LSAMEN( 2, C2, 'HE' ) ) THEN " the test works.



git-svn-id: https://icl.cs.utk.edu/svn/lapack-dev/lapack/trunk@1000 8a072113-8704-0410-8d35-dd094bca7971