Fix bug bug0088 reported by Mike Pont from NAG on the forum (see http://icl.cs.utk.edu/lapack-forum/viewtopic.php?f=13&t=2893)
This is related to the LAPACK-XBLAS routine: zherfsx.f

Here is what I did:
 - Introduce IINFO so that INFO is not overwritten
 - Use IGNORE_CWISE as suggested to prevent use of unitialize variable PARAMS
But this did not fix the problem reported.
INFO has new value in ZHERFSX (see description of INFO between ZHESVX and ZHESVXX)
This is set on line 634 (IF ( INFO .LE. N ) INFO = N + J) of zherfsx.f
And this is not handled by the testing LIN/zdrvhex.f
I just add  .AND. INFO.LE.N at line 638 to avoid raising an error when INFO = N + J

Please send feedback as I am not sure this is the best way to fix the issue.
I will commit other precision once fix approved.
Thanks
Julie



git-svn-id: https://icl.cs.utk.edu/svn/lapack-dev/lapack/trunk@1130 8a072113-8704-0410-8d35-dd094bca7971