corrected bug found by Tyrone Rees posted on email list:
"I think there's a bug in the routine dlaqp2.f in version 3.4.2 of LAPACK 
(and possibly earlier). Line 220 of the source is
   IF( I.LE.N ) THEN
whereas for the other versions the same line reads
   IF( I.LT.N ) THEN

This difference appears to make the routine dgeqp3 fall over when asked 
to do a factorization of the last two columns of a 10x5 matrix.  The NAG 
fortran compiler gives the error:
 Subscript out of range for assumed-size array A - Access to element 22 
but actual argument has only 20 elements
 Program terminated by fatal error

The single precision version runs without a hiccup.

Best wishes,

Tyrone"



git-svn-id: https://icl.cs.utk.edu/svn/lapack-dev/lapack/trunk@1385 8a072113-8704-0410-8d35-dd094bca7971