Ed Anderson Bug Fix - Jan 26 2015
Also in TESTING/EIG, I fixed two bugs in s/d chkhs; the diff of schkhs is:
790c790
<             CALL SLACPY( ' ', N, N, U, LDU, UZ, LDU )
---
>             CALL SLACPY( ' ', N, N, U, LDU, UZ, LDA )
831c831
<      $                 ABS( WI1( J )-WI3( J ) ) )
---
>      $                 ABS( WR1( J )-WR3( J ) ) )
In further testing, I was able to produce a test case where Test 8 fails comparing
the eigenvalues from xHSEQR(‘E’,’N’) with the eigenvalues from xHSEQR(‘S’,’V’).
The eigenvalues were the same, but in a different order.  It appears that xHSEQR
does some rearrangement of the Schur matrix and so we should not expect the
eigenvalues to be in the same order, although I am trying to confirm this 
with the authors.  In the meantime, I have modified xCHKEE and xCHKHS to 
compute the eigenvalues from xHSEQR(‘S’,’N’) in a separate array W2 (or 
WR2 and WI2) so that test 8 can compare the eigenvalues from xHSEQR(‘S’,’N’) 
with the eigenvalues from xHSEQR(‘S’,’V’).




git-svn-id: https://icl.cs.utk.edu/svn/lapack-dev/lapack/trunk@1522 8a072113-8704-0410-8d35-dd094bca7971