These are two fixes in [S,C,D,Z]-LAGGE.
(1) The routine was writing ZEROs out of bound in some specific cases. This is
corrected. Thanks to Yi Meng (CU Denver) for finding the bug.

(2) Quick exit when KL=KU=0 in this case, the user wants a diagonal matrix,
there is probably no reason to call this routine if one simply wants a diagonal
matrix, but that said we still ought to return a correct answer. A quick exit
makes the trick. The routine was not working otherwise.




git-svn-id: https://icl.cs.utk.edu/svn/lapack-dev/lapack/trunk@1552 8a072113-8704-0410-8d35-dd094bca7971